spotify/requirements.txt
spotipy>=2.23.0
python-dotenv>=0.19.0
